This source has now been adapted to compile without errors or warnings under both Visual Studio 2003 and Visual Studio 2008 (and probably Visual Studio 2005 as well, but I haven't tested this possibility).  There are separate .sln and .vcproj files to support either VS2003 or VS2008.  You should not get these mixed up as the 2008 solution files are not compatible with earlier versions of Visual Studio.
I have also put some initial project files to support editing this source from within Eclipse 3.4 (Ganymede), but this is purely experimental and I do not by any means believe that it will be compilable under Eclipse, at least not presently.  This project will always be a Microsoft MFC application, but it may be possible in the future to at least perform the development tasks in a more full featured development environment such as Eclipse.

While these files will compile under either VS2003 or VS2008, the application is only stable, presently, when built with VS2003.  I believe that this a linking problem with the C run time libraries which changed in VS2008, but hope to resolve this problem in the near future.  TinyCAD appears to build ok with VS2008, but it will crash randomly and solely due to problems in the C runtime, not from within TinyCAD, so you should only use a build of this source from VS2003.

This version also has some bug fixes and a few very minor new features.

Don


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@28 41062d70-b26d-460a-9302-f438a8930db8

+++ src/EditDlgPinEdit.cpp
@@ -153,7 +153,7 @@ void CEditDlgPinEdit::OnChange()
   static_cast<CDrawPin*>(getObject())->m_dir   = GetCheckedRadioButton(PINBOX_UP,PINBOX_RIGHT)-PINBOX_UP;
 
   int length = max(5,GetDlgItemInt(PINBOX_LENGTH));
-  int number_pos = min(length, GetDlgItemInt(PINBOX_PIN_POS ) );
+  int number_pos = min(length, (int) GetDlgItemInt(PINBOX_PIN_POS ) );
 
   static_cast<CDrawPin*>(getObject())->m_length = length;
   static_cast<CDrawPin*>(getObject())->m_number_pos = number_pos;
